requirements.txt
smbus2
i2c-tools
Mock-GPIO
Rpi-GPIO